Move existing shared module to be internal-only
