+++ src/internal/modules/routes.js
@@ -27,11 +27,11 @@ const viewLicences = require('./view-licences/routes')
 const gaugingStations = require('./gauging-stations/routes')
 const customers = require('./customers/routes')
 const notes = require('./notes/routes')
+const systemProxyRoutes = require('../../internal/modules/system-proxy/routes')
 
 // Shared routes
 const healthRoutes = require('../../shared/modules/health/routes')
 const kpiReporting = require('../../internal/modules/kpi-reporting/routes')
-const systemProxyRoutes = require('../../shared/modules/system-proxy/routes')
 
 // Internal search is being replaced by a new search page, but for some reason the existing internalSearch route
 // definition also includes additional routes relating to user maintenance. So we only remove the search route here
